+++ src/components/MenuItem/index.tsx
@@ -1,9 +1,0 @@
-export interface MenuItemProps {}
-
-export function MenuItem({}: MenuItemProps) {
-  return (
-    <div data-testid="menu-item">
-      <p>Menu Item</p>
-    </div>
-  );
-}